backend/requirements.txt
asgiref
Django
django-cors-headers #To fix cross origin request issue
djangorestframework
djangorestframework-simplejwt #Handle JWt authentication
PyJWT
pytz
sqlparse
psycopg2-binary #Postgres Database
python-dotenv